Extract element utility fns to `element_util.rs`
Makes `interface.rs` file easier to navigate.

Also extracted `enable_logs` to `logs.rs`.

This is a refactoring that I've extracted from #374,
where these functions are used by both the stateful
and the stateless interface and I wanted to avoid
any dependencies from `stateless_interface` on
`interface`.
